app/flask-app/requirements.txt
absl-py==1.3.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.6.5
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
autopep8==1.6.0
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.9.3
bitarray==2.9.2
bleach==5.0.1
cachetools==5.2.0
capstone==5.0.1
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
colored-traceback==0.3.0
cryptography==39.0.1
cvxopt==1.3.0.1
cycler==0.10.0
Cython==3.0.8
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
discord.py==1.7.3
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==0.9.1
factordb-pycli==1.3.0
fastapi==0.101.1
fastapi-sessions==0.3.2
fastjsonschema==2.16.1
Flask==2.2.5
Flask-Cors==3.0.10
flatbuffers==23.5.26
fonttools==4.38.0
frozendict==2.4.0
frozenlist==1.4.0
gast==0.4.0
gmpy2==2.1.5
google-api-core==2.11.1
google-api-python-client==2.100.0
google-auth==2.28.1
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
grpcio==1.50.0
gunicorn==20.1.0
h11==0.14.0
h5py==3.7.0
httplib2==0.22.0
idna==2.10
importlib-metadata==5.0.0
intervaltree==3.1.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isort==5.9.3
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
joblib==1.3.2
jsonschema==4.9.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
keras==2.15.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lapx==0.5.5
lazy-object-proxy==1.6.0
libclang==14.0.6
lxml==4.6.3
Mako==1.2.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
ml-dtypes==0.2.0
more-itertools==9.1.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.4
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
NBT==1.5.1
nest-asyncio==1.5.5
networkx==3.2.1
nltk==3.8.1
notebook==6.4.12
numpy==1.25.2
oauthlib==3.2.0
openai==0.28.0
opencv-python==4.8.0.76
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==23.1
pandas==1.4.3
pandocfilters==1.5.0
paramiko==3.4.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
plumbum==1.8.2
praw==7.3.0
prawcore==2.3.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==4.25.3
psutil==5.9.4
pure-eval==0.2.2
pwntools==4.12.0
py-cpuinfo==9.0.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.10.4
pydantic==2.3.0
pydantic_core==2.6.3
pydot==2.0.0
pyelftools==0.31
pygame==2.3.0
PyGetWindow==0.0.9
Pygments==2.12.0
pylint==2.9.6
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==3.0.9
PyPDF2==1.26.0
pyperclip==1.8.2
PyRect==0.1.4
pyrsistent==0.18.1
PyScreeze==0.1.27
pyserial==3.5
PySocks==1.7.1
pytesseract==0.3.10
pytholog==2.4.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
PyTweening==1.0.3
pytz==2022.1
pywin32==304
pywinpty==2.0.6
PyYAML==6.0.1
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
regex==2023.8.8
requests==2.25.1
requests-oauthlib==1.3.1
ROPGadget==7.4
rpyc==6.0.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.0
selenium==3.141.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.2.1
stack-data==0.3.0
starlette==0.27.0
sympy==1.11.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.1.1
terminado==0.15.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tinycss2==1.1.1
toml==0.10.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.66.1
traitlets==5.3.0
typing_extensions==4.7.1
ultralytics==8.1.5
unicorn==2.0.1.post1
unix-ar==0.2.1
update-checker==0.18.0
uritemplate==4.1.1
urllib3==1.26.5
uvicorn==0.23.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.3.3
widgetsnbextension==3.6.1
wrapt==1.14.1
xgboost==2.0.3
yarl==1.9.2
z3-solver==4.12.4.0
zipp==3.10.0
zstandard==0.22.0
